Improper drag attempts to a tableau pile are rejected. minor refactor
