+++ src/com/codecool/klondike/Pile.java
@@ -40,7 +40,6 @@ public class Pile extends Pane {
     }
 
     public int numOfCards() {
-        //TODO
         return cards.size();
     }
 
@@ -49,7 +48,6 @@ public class Pile extends Pane {
     }
 
     public void clear() {
-        //TODO
         cards.clear();
     }
